% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalind.R
\name{crossvalind}
\alias{crossvalind}
\title{Generate Cross-Validation Indices}
\usage{
crossvalind(N, kfold)
}
\arguments{
\item{N}{An integer specifying the total number of samples.}

\item{kfold}{An integer specifying the number of folds.}
}
\value{
A numeric vector of indices for cross-validation.
}
\description{
This function generates indices for k-fold cross-validation.
}
