% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadPredict.R
\name{roadPredict}
\alias{roadPredict}
\title{Predict Using a Fitted ROAD Model}
\usage{
roadPredict(xtest, fit, fitCV)
}
\arguments{
\item{xtest}{A numeric matrix of test data (n x p).}

\item{fit}{A fitted ROAD model object.}

\item{fitCV}{A cross-validated ROAD model object.}
}
\value{
A list containing:
  \item{class}{Predicted class labels for the optimal lambda.}
  \item{classAll}{Predicted class labels for all lambda values.}
}
\description{
This function predicts class labels for new data using a fitted ROAD model.
}
