% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat}
\alias{stat}
\title{Compute Statistics for ROAD Algorithm}
\usage{
stat(x, y)
}
\arguments{
\item{x}{A numeric matrix of predictors (n x p).}

\item{y}{A numeric vector of labels (0/1 or 1/2).}
}
\value{
A list containing:
  \item{x1}{Data for class 1.}
  \item{x2}{Data for class 2.}
  \item{n1}{Number of samples in class 1.}
  \item{n2}{Number of samples in class 2.}
  \item{mua}{Mean of the average of class means.}
  \item{mud}{Mean difference between classes.}
  \item{Sigma}{Pooled covariance matrix.}
}
\description{
This function computes necessary statistics (e.g., means, covariance) for the ROAD algorithm.
}
